+++ Backend/Database/Repositories/IThreadRepository.cs
@@ -6,5 +6,5 @@ public interface IThreadRepository : IRepository<Thread>
 {
     Task<IEnumerable<Thread>> GetByPageAsync(int page);
 
-    Task<IEnumerable<Thread>> GetByUserIdAsync(long userId, int page);
+    Task<IEnumerable<Thread>> GetByUserIdAsync(int userId, int page);
 }
